使用能确定长度的标准int类型

+++ User/bsp/a/bsp_delay.c
@@ -10,7 +10,7 @@
 /* Private variables -------------------------------------------------------- */
 /* Private function  -------------------------------------------------------- */
 /* Exported functions ------------------------------------------------------- */
-int BSP_Delay(uint32_t ms) {
+uint8_t BSP_Delay(uint32_t ms) {
   if (osKernelRunning()) {
     osDelay(ms);
   } else {
